Express-router
Separando lo que es la app en si ósea el servidor, de las request que solicita el usuario al servidor para tener todo mejor estructurado en diferentes archivos que se comunican entre si.
